added histograms, refactored functions into separate files
